Finished testing clock stuff
