+++ lab0/CMakeLists.txt
@@ -9,12 +9,15 @@ file(GLOB_RECURSE SOURCES_C "src/*.c")
 
 set(SOURCES ${SOURCES_CPP} ${SOURCES_C})
 
+set(CMAKE_C_COMPILER clang)
+set(CMAKE_CXX_COMPILER clang++)
+
 # Set the include directories.
 include_directories("./include")
 
-# Set compiler flags
-set(COMMON_FLAGS "-Wall -pedantic -fpermissive")
-set(CMAKE_C_FLAGS "${COMMON_FLAGS}")
+# Set compiler flags and linker flags
+SET(COMPILE_FLAGS "-Wall -pedantic -fpermissive ")
+SET(LINK_FLAGS    "-lrt")
 
 # Set the executable.
 add_executable(${CMAKE_PROJECT_NAME} ${SOURCES})
@@ -25,10 +28,6 @@ set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib")
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib")
 
 # OS specific options and libraries
-if(WIN32)
-  # Disable warning 4996.
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /wd4996")
-else()
-  # Enable all pedantic warnings.
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x ${COMMON_FLAGS}")
-endif()
+SET(CMAKE_CXX_FLAGS  "${CMAKE_CXX_FLAGS} ${COMPILE_FLAGS}")
+SET(CMAKE_EXE_LINKER_FLAGS  "${CMAKE_EXE_LINKER_FLAGS} ${LINK_FLAGS}")
+
